formatting and comments
