+++ CMakeLists.txt
@@ -14,7 +14,7 @@ include(FindPkgConfig)
 include(GNUInstallDirs)
 
 set(VALHALLA_SOURCE_DIR ${CMAKE_CURRENT_SOURCE_DIR})
-set(VALHALLA_BUILD_DIR ${CMAKE_CURRENT_BINARY_DIR})
+set(VALHALLA_BUILD_DIR ${CMxAKE_CURRENT_BINARY_DIR})
 list(INSERT CMAKE_MODULE_PATH 0 ${VALHALLA_SOURCE_DIR}/cmake)
 
 set(CMAKE_CXX_STANDARD 14 CACHE STRING "C++ language version to use (default is 14)")
