+++ src/mjolnir/valhalla_ingest_transit.cc
@@ -3,7 +3,7 @@
 
 int main(int argc, char** argv) {
   if (argc < 2) {
-    std::cerr << "Usage: " << std::string(argv[0]) << "Error" << std::endl;
+    std::cerr << "Usage: " << std::string(argv[0]) << " Error" << std::endl;
     return 1;
   }
